requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
Columnar==1.4.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.1
openai==1.44.1
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.91
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soupsieve==2.6
stack-data==0.6.3
string-color==1.2.3
tinycss2==1.3.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-requests==2.32.0.20240907
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
